requirements.txt
discord.py
python-dotenv
pandas
numpy
xgboost
scikit-learn
joblib
schedule
